refactor: 포인트 관련 필드 long 타입으로 변경
